fix(chat): gate org-file refs on provider context
Lift the org-file open behavior into an OrgFileOpenProvider mounted on the
agent shell, consumed by the markdown code/link overrides via context.

- No dead clicks off-chat: MemoizedMarkdown also renders in the Library's
  file/skill previews (/$org/files), which has no main panel and no `?main=`
  key, so `?main=library-file:` was silently dropped while the chip still
  looked clickable. Without the provider the ref now renders as plain code.
- Perf: useNavigate/useIsMobile/useParams run once in the provider instead of
  per inline <code>/<a> across every message, file preview and PR tab.
